# puzzle_backend/swagger/swagger.yml
swagger: "2.0"
info:
  title: "Harmony Puzzle"
  description: >
    Harmony Puzzle is a simple yet addictive blockchain-based puzzle game.
  version: "0.0.1"
host: "harmony-puzzle-ek.appspot.com"
basePath: /v1
schemes:
  - https
paths:
  "/reg":
    post:
      operationId: "postReg"
      summary: FE & ContentOS calls this API when the Harmony game is loaded.
      consumes:
        - application/json
      parameters:
        - in: query
          name: email
          type: string
          description: user's email
          required: true
      produces:
        - application/json
      responses:
        "201":
          description: A new user.  FE should instruct the user to check email.
          headers:
            "Access-Control-Allow-Origin":
              type: string
              description: CORS origin bypass. TODO ek – remove this.
          schema:
            type: object
            properties:
              email:
                type: string
              account:
                type: string
        "200":
          description: An existing user.
          schema:
            type: object
            properties:
              email:
                type: string
              account:
                type: string
        "401":
          description: The given token is invalid.
          schema:
            type: object
            properties:
              msg:
                description: The error message.
                type: string
  "/play":
    post:
      operationId: "postPlay"
      summary: FE calls this to start a new game.
      consumes:
        - application/json
      parameters:
        - in: query
          name: accountKey
          type: string
          description: user's account private key, hex-encoded
          required: true
        - in: query
          name: stake
          type: number
          description: user's bet, in wei (divide by 10^18 to get HRX)
          required: true
      responses:
        "201":
          description: A new game has been started.
          schema:
            type: object
            properties:
              txid:
                type: string
                description: staking txid
  "/finish":
    post:
      operationId: "postFinish"
      summary: FE calls this to post a level
      consumes:
        - application/json
      parameters:
        - in: query
          name: accountKey
          type: string
          description: user's account private key, hex-encoded
          required: true
        - in: query
          name: txid
          type: string
          description: the game ID (staking transaction ID) returned by POST /play
          required: true
        - in: query
          name: level
          type: integer
          description: level number (1-based)
          required: true
        - in: query
          name: height
          type: integer
          description: the solution's height (the same number found in all squares)
        - in: query
          name: sequence
          type: string
          description: user's moves from first to last; [udlr]* in regex
          required: true
        - in: query
          name: last_pos
          type: integer
          description: where the cursor was after completing the last move in sequence, in telephone keypad notation (1-9)
          required: true
      responses:
        "200":
          description: The level has been accepted.
          schema:
            type: object
            properties:
              reward:
                type: number
                description: reward amount, in wei (divide by 10^18 to get HRX)